Move Broken Giant's Knife fix hook to its own file
